requirements.txt
# DAP - 数据处理审计智能体
# 核心依赖包

# 数据处理
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Web框架和API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# HTTP请求
requests>=2.31.0

# 配置文件处理
PyYAML>=6.0.1

# 日期时间处理
python-dateutil>=2.8.2

# 文件类型检测
python-magic-bin>=0.4.14  # Windows版本
# python-magic>=0.4.27    # Linux/Mac版本

# GUI界面
tkinterdnd2>=0.3.0

# AI/机器学习（可选）
# openai>=1.0.0  # 如果使用OpenAI API
# ollama>=0.1.0  # 如果使用本地LLM

# 日志
colorlog>=6.7.0

# 进度条
tqdm>=4.66.0

# 工具包 - built-in modules removed

# HTTP CORS支持
fastapi-cors>=0.0.6

# 数据可视化（可选）
matplotlib>=3.7.0
seaborn>=0.12.0

# 报告生成（可选）
reportlab>=4.0.0

# 压缩文件处理
rarfile>=4.0  # RAR文件支持（可选）

# 统计分析（可选）
scipy>=1.11.0
scikit-learn>=1.3.0

# 开发和测试（可选）
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# 性能监控（可选）
memory-profiler>=0.61.0
psutil>=5.9.0